backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
pyjwt
supabase
